Tests/LibWeb/Text/expected/geometry/getClientRects-wrapped-ltr-span-in-rtl-block.txt

rects: 2
wrap[0]: 17,5 38x10
wrap[1]: 23,25 32x10
